Move "resBrut" to results.casTypes.items
