+++ .eslintrc.json
@@ -28,6 +28,7 @@
     }
   },
   "rules": {
+    "no-case-declarations": 0,
     "local/no-material-root-imports": 2,
     "sort-keys": [2, "asc"],
     "array-bracket-newline": ["error", { "multiline": true }],
