+++ components/cartes-impact/index.js
@@ -5,7 +5,6 @@ import ImpactComponent from "./impact-component";
 const mapStateToProps = ({
   casTypes,
   display,
-  resBrut,
   token,
   totalPop,
 }) => {
@@ -14,7 +13,6 @@ const mapStateToProps = ({
     casTypes,
     isInformationPanelVisible: display.isInformationPanelVisible,
     isUserLogged,
-    resBrut,
     totalPop,
   };
 };
