+++ components/cartes-impact/simple-card/index.js
@@ -3,18 +3,18 @@ import { connect } from "react-redux";
 import { removeCasType, showEditCasTypesPopin, simulateCasTypes } from "../../../redux/actions";
 import SimpleCard from "./simple-card-component";
 
-const mapStateToProps = ({ casTypes, resBrut, results }, { index }) => {
-  const isLoading = results.casTypes.isFetching;
+const mapStateToProps = ({ casTypes, results }, { index }) => {
+  const { isFetching, items } = results.casTypes;
   const { name } = casTypes[index];
   const descCasType = casTypes[index];
   const resultats = {
-    apres: resBrut.apres[index],
-    avant: resBrut.avant[index],
-    plf: resBrut.plf ? resBrut.plf[index] : null,
+    apres: items.apres[index],
+    avant: items.avant[index],
+    plf: items.plf ? items.plf[index] : null,
   };
   return {
     descCasType,
-    isLoading,
+    isLoading: isFetching,
     name,
     resultats,
   };
